requirements.txt
google-generativeai
streamlit
fastapi
pydantic
uvicorn
PyPDF2
pypdf
python-docx
beautifulsoup4
python-dotenv
streamlit-authenticator
PyYAML
